+++ +gams/+transfer/+symbol/Equation.m
@@ -289,6 +289,7 @@ classdef Equation < gams.transfer.symbol.Symbol
             end
 
             symbol.copyFrom(obj);
+            symbol.def.switchContainer(destination);
         end
 
     end
@@ -316,7 +317,7 @@ classdef Equation < gams.transfer.symbol.Symbol
             % and some statistics.
 
             symbols = gams.transfer.utils.validate_cell('symbols', 1, symbols, ...
-                {'gams.transfer.symbol.Equation'}, 1);
+                {'gams.transfer.symbol.Equation'}, 1, -1);
 
             descr = struct();
             descr.name = cell(numel(symbols), 1);