+++ +gams/+transfer/+symbol/Set.m
@@ -266,6 +266,7 @@ classdef Set < gams.transfer.symbol.Symbol
             end
 
             symbol.copyFrom(obj);
+            symbol.def.switchContainer(destination);
         end
 
     end
@@ -293,7 +294,7 @@ classdef Set < gams.transfer.symbol.Symbol
             % and some statistics.
 
             symbols = gams.transfer.utils.validate_cell('symbols', 1, symbols, ...
-                {'gams.transfer.symbol.Set'}, 1);
+                {'gams.transfer.symbol.Set'}, 1, -1);
 
             descr = struct();
             descr.name = cell(numel(symbols), 1);